+++ HISTORY.rst
@@ -3,6 +3,17 @@
 Release History
 ---------------
 
+3.3.0 (2023-01-06)
+++++++++++++++++++
+
+**Features and Improvements**
+
+- Added support for inserting metadata into an existing multi-value metadata
+  field. It differes from ``ia metadata <id> --modify collection[0]:foo`` in
+  that it does not clobber. For example,
+  ``ia metadata <id> --insert collection[0]:foo`` will insert ``foo`` as the
+  first collection, it will not clobber.
+
 3.2.0 (2023-01-06)
 ++++++++++++++++++